[http] Make chunk_remaining also off_t
off_t is usually 8 bytes long. HTTP/1.1 does not say
that chunks cannot be bigger than int.

+++ src/protocol/http/http.h
@@ -30,7 +30,7 @@ struct http_connection_info {
 
 	int close;
 	off_t length;
-	int chunk_remaining;
+	off_t chunk_remaining;
 	int code;
 
 	struct http_post post;
